translated opentdb api wrapper from node to js
